Make section size 'small' the default size, Move files to folders.
